reorder local variables

+++ IPv4/Outbound/WindowsServices.ps1
@@ -28,9 +28,6 @@ SOFTWARE.
 . $PSScriptRoot\..\IPSetup.ps1
 Import-Module -Name $PSScriptRoot\..\..\FirewallModule
 
-# Ask user if he wants to load these rules
-if (!(Approve-Execute)) { exit }
-
 #
 # Setup local variables:
 #
@@ -40,6 +37,10 @@ $Profile = "Private, Public"
 [string[]] $ExtensionAccounts = @("NT AUTHORITY\SYSTEM", "NT AUTHORITY\LOCAL SERVICE", "NT AUTHORITY\NETWORK SERVICE")
 $ExtensionAccounts += $UserAccounts
 $ExtensionUsers = (Get-AccountSDDL $ExtensionAccounts)
+
+# Ask user if he wants to load these rules
+Update-Context $IPVersion $Direction $Group
+if (!(Approve-Execute)) { exit }
 
 # First remove all existing rules matching group
 Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direction -ErrorAction SilentlyContinue
